Clean up cross-platform diagnostic
